+++ src/main/kotlin/domain/state/running/Running.kt
@@ -1,5 +1,6 @@
-package domain.state
+package domain.state.running
 
+import domain.state.State
 import domain.stone.Board
 import domain.stone.Stone
 import domain.stone.StoneType
