+++ src/test/kotlin/domain/state/end/EndTest.kt
@@ -1,6 +1,6 @@
-package domain.stone
+package domain.state.end
 
-import domain.state.End
+import domain.stone.StoneType
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
